Move NoteMenu and related components to a separate file
